requirements.txt
https://github.com/celery/celery/tarball/master  # latest version from pip doesn't work with python 3.7
django==2.1.5
factory_boy==2.11.1
psycopg2-binary==2.7.6.1
redis==3.0.1